+++ src/collections.ts
@@ -26,7 +26,7 @@ import { COLLECTION_NOT_FOUND, DOCUMENT_NOT_FOUND } from "./lib/codes.js";
  * @param collection - A value that might be a collection.
  */
 export function isArangoCollection(
-  collection: any,
+  collection: any
 ): collection is ArangoCollection {
   return Boolean(collection && collection.isArangoCollection);
 }
@@ -38,7 +38,7 @@ export function isArangoCollection(
  * @param collection - Collection name or {@link ArangoCollection} object.
  */
 export function collectionToString(
-  collection: string | ArangoCollection,
+  collection: string | ArangoCollection
 ): string {
   if (isArangoCollection(collection)) {
     return String(collection.name);
@@ -663,7 +663,7 @@ export interface DocumentCollection<
   create(
     options?: CreateCollectionOptions & {
       type?: CollectionType;
-    },
+    }
   ): Promise<
     connection.ArangoApiResponse<CollectionDescription & CollectionProperties>
   >;
@@ -694,7 +694,7 @@ export interface DocumentCollection<
    * ```
    */
   properties(
-    properties: CollectionPropertiesOptions,
+    properties: CollectionPropertiesOptions
   ): Promise<
     connection.ArangoApiResponse<CollectionDescription & CollectionProperties>
   >;
@@ -745,7 +745,7 @@ export interface DocumentCollection<
    * ```
    */
   figures(
-    details?: boolean,
+    details?: boolean
   ): Promise<
     connection.ArangoApiResponse<
       CollectionDescription &
@@ -782,7 +782,7 @@ export interface DocumentCollection<
    * ```
    */
   checksum(
-    options?: CollectionChecksumOptions,
+    options?: CollectionChecksumOptions
   ): Promise<
     connection.ArangoApiResponse<
       CollectionDescription & { revision: string; checksum: string }
@@ -811,7 +811,7 @@ export interface DocumentCollection<
    * ```
    */
   rename(
-    newName: string,
+    newName: string
   ): Promise<connection.ArangoApiResponse<CollectionDescription>>;
   /**
    * Deletes all documents in the collection.
@@ -826,7 +826,7 @@ export interface DocumentCollection<
    * ```
    */
   truncate(
-    options?: TruncateCollectionOptions,
+    options?: TruncateCollectionOptions
   ): Promise<connection.ArangoApiResponse<CollectionDescription>>;
   /**
    * Deletes the collection from the database.
@@ -842,7 +842,7 @@ export interface DocumentCollection<
    * ```
    */
   drop(
-    options?: DropCollectionOptions,
+    options?: DropCollectionOptions
   ): Promise<connection.ArangoApiResponse<{ id: string }>>;
   /**
    * Triggers compaction for a collection.
@@ -872,7 +872,7 @@ export interface DocumentCollection<
    * ```
    */
   getResponsibleShard(
-    document: Partial<documents.Document<EntryResultType>>,
+    document: Partial<documents.Document<EntryResultType>>
   ): Promise<string>;
   /**
    * Derives a document `_id` from the given selector for this collection.
@@ -930,7 +930,7 @@ export interface DocumentCollection<
    */
   documentExists(
     selector: documents.DocumentSelector,
-    options?: documents.DocumentExistsOptions,
+    options?: documents.DocumentExistsOptions
   ): Promise<boolean>;
   /**
    * Retrieves the document matching the given key or id.
@@ -968,7 +968,7 @@ export interface DocumentCollection<
    */
   document(
     selector: documents.DocumentSelector,
-    options?: documents.ReadDocumentOptions,
+    options?: documents.ReadDocumentOptions
   ): Promise<documents.Document<EntryResultType>>;
   /**
    * Retrieves the document matching the given key or id.
@@ -1007,7 +1007,7 @@ export interface DocumentCollection<
    */
   document(
     selector: documents.DocumentSelector,
-    graceful: boolean,
+    graceful: boolean
   ): Promise<documents.Document<EntryResultType>>;
   /**
    * Retrieves the documents matching the given key or id values.
@@ -1033,7 +1033,7 @@ export interface DocumentCollection<
    */
   documents(
     selectors: (string | documents.ObjectWithDocumentKey)[],
-    options?: documents.BulkReadDocumentsOptions,
+    options?: documents.BulkReadDocumentsOptions
   ): Promise<documents.Document<EntryResultType>[]>;
   /**
    * Inserts a new document with the given `data` into the collection.
@@ -1054,7 +1054,7 @@ export interface DocumentCollection<
    */
   save(
     data: documents.DocumentData<EntryInputType>,
-    options?: documents.InsertDocumentOptions,
+    options?: documents.InsertDocumentOptions
   ): Promise<
     documents.DocumentOperationMetadata & {
       new?: documents.Document<EntryResultType>;
@@ -1084,7 +1084,7 @@ export interface DocumentCollection<
    */
   saveAll(
     data: Array<documents.DocumentData<EntryInputType>>,
-    options?: documents.InsertDocumentOptions,
+    options?: documents.InsertDocumentOptions
   ): Promise<
     Array<
       | (documents.DocumentOperationMetadata & {
@@ -1121,7 +1121,7 @@ export interface DocumentCollection<
   replace(
     selector: documents.DocumentSelector,
     newData: documents.DocumentData<EntryInputType>,
-    options?: documents.ReplaceDocumentOptions,
+    options?: documents.ReplaceDocumentOptions
   ): Promise<
     documents.DocumentOperationMetadata & {
       new?: documents.Document<EntryResultType>;
@@ -1157,7 +1157,7 @@ export interface DocumentCollection<
       documents.DocumentData<EntryInputType> &
         ({ _key: string } | { _id: string })
     >,
-    options?: Omit<documents.ReplaceDocumentOptions, "ifMatch">,
+    options?: Omit<documents.ReplaceDocumentOptions, "ifMatch">
   ): Promise<
     Array<
       | (documents.DocumentOperationMetadata & {
@@ -1194,7 +1194,7 @@ export interface DocumentCollection<
   update(
     selector: documents.DocumentSelector,
     newData: documents.Patch<documents.DocumentData<EntryInputType>>,
-    options?: documents.UpdateDocumentOptions,
+    options?: documents.UpdateDocumentOptions
   ): Promise<
     documents.DocumentOperationMetadata & {
       new?: documents.Document<EntryResultType>;
@@ -1230,7 +1230,7 @@ export interface DocumentCollection<
       documents.Patch<documents.DocumentData<EntryInputType>> &
         ({ _key: string } | { _id: string })
     >,
-    options?: Omit<documents.UpdateDocumentOptions, "ifMatch">,
+    options?: Omit<documents.UpdateDocumentOptions, "ifMatch">
   ): Promise<
     Array<
       | (documents.DocumentOperationMetadata & {
@@ -1269,7 +1269,7 @@ export interface DocumentCollection<
    */
   remove(
     selector: documents.DocumentSelector,
-    options?: documents.RemoveDocumentOptions,
+    options?: documents.RemoveDocumentOptions
   ): Promise<
     documents.DocumentMetadata & { old?: documents.Document<EntryResultType> }
   >;
@@ -1293,7 +1293,7 @@ export interface DocumentCollection<
    */
   removeAll(
     selectors: (string | documents.ObjectWithDocumentKey)[],
-    options?: Omit<documents.RemoveDocumentOptions, "ifMatch">,
+    options?: Omit<documents.RemoveDocumentOptions, "ifMatch">
   ): Promise<
     Array<
       | (documents.DocumentMetadata & {
@@ -1323,7 +1323,7 @@ export interface DocumentCollection<
    */
   import(
     data: documents.DocumentData<EntryInputType>[],
-    options?: documents.ImportDocumentsOptions,
+    options?: documents.ImportDocumentsOptions
   ): Promise<documents.ImportDocumentsResult>;
   /**
    * Bulk imports the given `data` into the collection.
@@ -1349,7 +1349,7 @@ export interface DocumentCollection<
    */
   import(
     data: any[][],
-    options?: documents.ImportDocumentsOptions,
+    options?: documents.ImportDocumentsOptions
   ): Promise<documents.ImportDocumentsResult>;
   /**
    * Bulk imports the given `data` into the collection.
@@ -1411,7 +1411,7 @@ export interface DocumentCollection<
     data: Buffer | Blob | string,
     options?: documents.ImportDocumentsOptions & {
       type?: "documents" | "list" | "auto";
-    },
+    }
   ): Promise<documents.ImportDocumentsResult>;
   //#endregion
 
@@ -1455,7 +1455,7 @@ export interface DocumentCollection<
       | indexes.IndexDescription
       | indexes.HiddenIndexDescription = indexes.IndexDescription,
   >(
-    options?: indexes.ListIndexesOptions,
+    options?: indexes.ListIndexesOptions
   ): Promise<IndexType[]>;
   /**
    * Returns an index description by name or `id` if it exists.
@@ -1489,7 +1489,7 @@ export interface DocumentCollection<
    * ```
    */
   ensureIndex(
-    options: indexes.EnsurePersistentIndexOptions,
+    options: indexes.EnsurePersistentIndexOptions
   ): Promise<
     connection.ArangoApiResponse<
       indexes.PersistentIndexDescription & { isNewlyCreated: boolean }
@@ -1525,7 +1525,7 @@ export interface DocumentCollection<
    * ```
    */
   ensureIndex(
-    options: indexes.EnsureTtlIndexOptions,
+    options: indexes.EnsureTtlIndexOptions
   ): Promise<
     connection.ArangoApiResponse<
       indexes.TtlIndexDescription & { isNewlyCreated: boolean }
@@ -1550,10 +1550,37 @@ export interface DocumentCollection<
    * ```
    */
   ensureIndex(
-    options: indexes.EnsureMdiIndexOptions,
+    options: indexes.EnsureMdiIndexOptions
   ): Promise<
     connection.ArangoApiResponse<
       indexes.MdiIndexDescription & { isNewlyCreated: boolean }
+    >
+  >;
+  /**
+   * Creates a prefixed multi-dimensional index on the collection if it does
+   * not already exist.
+   *
+   * @param options - Options for creating the prefixed multi-dimensional index.
+   *
+   * @example
+   * ```js
+   * const db = new Database();
+   * const collection = db.collection("some-points");
+   * // Create a multi-dimensional index for the attributes x, y and z
+   * await collection.ensureIndex({
+   *   type: "mdi-prefixed",
+   *   fields: ["x", "y", "z"],
+   *   prefixFields: ["x"],
+   *   fieldValueTypes: "double"
+   * });
+   * ```
+   * ```
+   */
+  ensureIndex(
+    options: indexes.EnsureMdiPrefixedIndexOptions
+  ): Promise<
+    connection.ArangoApiResponse<
+      indexes.MdiPrefixedIndexDescription & { isNewlyCreated: boolean }
     >
   >;
   /**
@@ -1574,7 +1601,7 @@ export interface DocumentCollection<
    * ```
    */
   ensureIndex(
-    options: indexes.EnsureGeoIndexOptions,
+    options: indexes.EnsureGeoIndexOptions
   ): Promise<
     connection.ArangoApiResponse<
       indexes.GeoIndexDescription & { isNewlyCreated: boolean }
@@ -1597,7 +1624,7 @@ export interface DocumentCollection<
    * ```
    */
   ensureIndex(
-    options: indexes.EnsureInvertedIndexOptions,
+    options: indexes.EnsureInvertedIndexOptions
   ): Promise<
     connection.ArangoApiResponse<
       indexes.InvertedIndexDescription & { isNewlyCreated: boolean }
@@ -1622,7 +1649,7 @@ export interface DocumentCollection<
    * ```
    */
   ensureIndex(
-    options: indexes.EnsureIndexOptions,
+    options: indexes.EnsureIndexOptions
   ): Promise<
     connection.ArangoApiResponse<
       indexes.IndexDescription & { isNewlyCreated: boolean }
@@ -1642,7 +1669,7 @@ export interface DocumentCollection<
    * ```
    */
   dropIndex(
-    selector: indexes.IndexSelector,
+    selector: indexes.IndexSelector
   ): Promise<connection.ArangoApiResponse<{ id: string }>>;
   //#endregion
 }
@@ -1717,7 +1744,7 @@ export interface EdgeCollection<
    */
   document(
     selector: documents.DocumentSelector,
-    options?: documents.ReadDocumentOptions,
+    options?: documents.ReadDocumentOptions
   ): Promise<documents.Edge<EntryResultType>>;
   /**
    * Retrieves the document matching the given key or id.
@@ -1756,7 +1783,7 @@ export interface EdgeCollection<
    */
   document(
     selector: documents.DocumentSelector,
-    graceful: boolean,
+    graceful: boolean
   ): Promise<documents.Edge<EntryResultType>>;
   /**
    * Retrieves the documents matching the given key or id values.
@@ -1782,7 +1809,7 @@ export interface EdgeCollection<
    */
   documents(
     selectors: (string | documents.ObjectWithDocumentKey)[],
-    options?: documents.BulkReadDocumentsOptions,
+    options?: documents.BulkReadDocumentsOptions
   ): Promise<documents.Edge<EntryResultType>[]>;
   /**
    * Inserts a new document with the given `data` into the collection.
@@ -1802,7 +1829,7 @@ export interface EdgeCollection<
    */
   save(
     data: documents.EdgeData<EntryInputType>,
-    options?: documents.InsertDocumentOptions,
+    options?: documents.InsertDocumentOptions
   ): Promise<
     documents.DocumentOperationMetadata & {
       new?: documents.Edge<EntryResultType>;
@@ -1830,7 +1857,7 @@ export interface EdgeCollection<
    */
   saveAll(
     data: Array<documents.EdgeData<EntryInputType>>,
-    options?: documents.InsertDocumentOptions,
+    options?: documents.InsertDocumentOptions
   ): Promise<
     Array<
       | (documents.DocumentOperationMetadata & {
@@ -1875,7 +1902,7 @@ export interface EdgeCollection<
   replace(
     selector: documents.DocumentSelector,
     newData: documents.DocumentData<EntryInputType>,
-    options?: documents.ReplaceDocumentOptions,
+    options?: documents.ReplaceDocumentOptions
   ): Promise<
     documents.DocumentOperationMetadata & {
       new?: documents.Edge<EntryResultType>;
@@ -1927,7 +1954,7 @@ export interface EdgeCollection<
       documents.DocumentData<EntryInputType> &
         ({ _key: string } | { _id: string })
     >,
-    options?: documents.ReplaceDocumentOptions,
+    options?: documents.ReplaceDocumentOptions
   ): Promise<
     Array<
       | (documents.DocumentOperationMetadata & {
@@ -1972,7 +1999,7 @@ export interface EdgeCollection<
   update(
     selector: documents.DocumentSelector,
     newData: documents.Patch<documents.DocumentData<EntryInputType>>,
-    options?: documents.UpdateDocumentOptions,
+    options?: documents.UpdateDocumentOptions
   ): Promise<
     documents.DocumentOperationMetadata & {
       new?: documents.Edge<EntryResultType>;
@@ -2022,7 +2049,7 @@ export interface EdgeCollection<
       documents.Patch<documents.DocumentData<EntryInputType>> &
         ({ _key: string } | { _id: string })
     >,
-    options?: documents.UpdateDocumentOptions,
+    options?: documents.UpdateDocumentOptions
   ): Promise<
     Array<
       | (documents.DocumentOperationMetadata & {
@@ -2053,7 +2080,7 @@ export interface EdgeCollection<
    */
   remove(
     selector: documents.DocumentSelector,
-    options?: documents.RemoveDocumentOptions,
+    options?: documents.RemoveDocumentOptions
   ): Promise<
     documents.DocumentMetadata & { old?: documents.Edge<EntryResultType> }
   >;
@@ -2077,7 +2104,7 @@ export interface EdgeCollection<
    */
   removeAll(
     selectors: documents.DocumentSelector[],
-    options?: documents.RemoveDocumentOptions,
+    options?: documents.RemoveDocumentOptions
   ): Promise<
     Array<
       | (documents.DocumentMetadata & { old?: documents.Edge<EntryResultType> })
@@ -2104,7 +2131,7 @@ export interface EdgeCollection<
    */
   import(
     data: documents.EdgeData<EntryInputType>[],
-    options?: documents.ImportDocumentsOptions,
+    options?: documents.ImportDocumentsOptions
   ): Promise<documents.ImportDocumentsResult>;
   /**
    * Bulk imports the given `data` into the collection.
@@ -2129,7 +2156,7 @@ export interface EdgeCollection<
    */
   import(
     data: any[][],
-    options?: documents.ImportDocumentsOptions,
+    options?: documents.ImportDocumentsOptions
   ): Promise<documents.ImportDocumentsResult>;
   /**
    * Bulk imports the given `data` into the collection.
@@ -2188,7 +2215,7 @@ export interface EdgeCollection<
     data: Buffer | Blob | string,
     options?: documents.ImportDocumentsOptions & {
       type?: "documents" | "list" | "auto";
-    },
+    }
   ): Promise<documents.ImportDocumentsResult>;
   //#endregion
 
@@ -2220,7 +2247,7 @@ export interface EdgeCollection<
    */
   edges(
     selector: documents.DocumentSelector,
-    options?: documents.DocumentEdgesOptions,
+    options?: documents.DocumentEdgesOptions
   ): Promise<
     connection.ArangoApiResponse<documents.DocumentEdgesResult<EntryResultType>>
   >;
@@ -2251,7 +2278,7 @@ export interface EdgeCollection<
    */
   inEdges(
     selector: documents.DocumentSelector,
-    options?: documents.DocumentEdgesOptions,
+    options?: documents.DocumentEdgesOptions
   ): Promise<
     connection.ArangoApiResponse<documents.DocumentEdgesResult<EntryResultType>>
   >;
@@ -2282,7 +2309,7 @@ export interface EdgeCollection<
    */
   outEdges(
     selector: documents.DocumentSelector,
-    options?: documents.DocumentEdgesOptions,
+    options?: documents.DocumentEdgesOptions
   ): Promise<
     connection.ArangoApiResponse<documents.DocumentEdgesResult<EntryResultType>>
   >;
@@ -2347,7 +2374,7 @@ export class Collection<
   create(
     options: CreateCollectionOptions & {
       type?: CollectionType;
-    } = {},
+    } = {}
   ) {
     const {
       waitForSyncReplication = undefined,
@@ -2390,7 +2417,7 @@ export class Collection<
   }
 
   properties(
-    properties?: CollectionPropertiesOptions,
+    properties?: CollectionPropertiesOptions
   ): Promise<
     connection.ArangoApiResponse<CollectionDescription & CollectionProperties>
   > {
@@ -2421,15 +2448,15 @@ export class Collection<
       {
         method: "PUT",
         pathname: `/_api/collection/${encodeURIComponent(
-          this._name,
+          this._name
         )}/recalculateCount`,
       },
-      (res) => res.parsedBody.result,
+      (res) => res.parsedBody.result
     );
   }
 
   figures(
-    details = false,
+    details = false
   ): Promise<
     connection.ArangoApiResponse<
       CollectionDescription &
@@ -2453,7 +2480,7 @@ export class Collection<
   }
 
   checksum(
-    options?: CollectionChecksumOptions,
+    options?: CollectionChecksumOptions
   ): Promise<
     connection.ArangoApiResponse<
       CollectionDescription & { revision: string; checksum: string }
@@ -2472,7 +2499,7 @@ export class Collection<
   }
 
   truncate(
-    options?: TruncateCollectionOptions,
+    options?: TruncateCollectionOptions
   ): Promise<connection.ArangoApiResponse<CollectionDescription>> {
     return this._db.request({
       method: "PUT",
@@ -2499,17 +2526,17 @@ export class Collection<
 
   //#region Document operations
   getResponsibleShard(
-    document: Partial<documents.Document<EntryResultType>>,
+    document: Partial<documents.Document<EntryResultType>>
   ): Promise<string> {
     return this._db.request(
       {
         method: "PUT",
         pathname: `/_api/collection/${encodeURIComponent(
-          this._name,
+          this._name
         )}/responsibleShard`,
         body: document,
       },
-      (res) => res.parsedBody.shardId,
+      (res) => res.parsedBody.shardId
     );
   }
 
@@ -2519,7 +2546,7 @@ export class Collection<
 
   async documentExists(
     selector: documents.DocumentSelector,
-    options: documents.DocumentExistsOptions = {},
+    options: documents.DocumentExistsOptions = {}
   ): Promise<boolean> {
     const { ifMatch = undefined, ifNoneMatch = undefined } = options;
     const headers = {} as Record<string, string>;
@@ -2530,7 +2557,7 @@ export class Collection<
         {
           method: "HEAD",
           pathname: `/_api/document/${encodeURI(
-            documents._documentHandle(selector, this._name),
+            documents._documentHandle(selector, this._name)
           )}`,
           headers,
         },
@@ -2539,7 +2566,7 @@ export class Collection<
             throw new errors.HttpError(res);
           }
           return true;
-        },
+        }
       );
     } catch (err: any) {
       if (err.code === 404) {
@@ -2551,7 +2578,7 @@ export class Collection<
 
   documents(
     selectors: (string | documents.ObjectWithDocumentKey)[],
-    options: documents.BulkReadDocumentsOptions = {},
+    options: documents.BulkReadDocumentsOptions = {}
   ) {
     const { allowDirtyRead = undefined } = options;
     return this._db.request({
@@ -2565,7 +2592,7 @@ export class Collection<
 
   async document(
     selector: documents.DocumentSelector,
-    options: boolean | documents.ReadDocumentOptions = {},
+    options: boolean | documents.ReadDocumentOptions = {}
   ) {
     if (typeof options === "boolean") {
       options = { graceful: options };
@@ -2582,7 +2609,7 @@ export class Collection<
     const result = this._db.request(
       {
         pathname: `/_api/document/${encodeURI(
-          documents._documentHandle(selector, this._name),
+          documents._documentHandle(selector, this._name)
         )}`,
         headers,
         allowDirtyRead,
@@ -2592,7 +2619,7 @@ export class Collection<
           throw new errors.HttpError(res);
         }
         return res.parsedBody;
-      },
+      }
     );
     if (!graceful) return result;
     try {
@@ -2607,7 +2634,7 @@ export class Collection<
 
   save(
     data: documents.DocumentData<EntryInputType>,
-    options?: documents.InsertDocumentOptions,
+    options?: documents.InsertDocumentOptions
   ) {
     return this._db.request(
       {
@@ -2616,13 +2643,13 @@ export class Collection<
         body: data,
         search: options,
       },
-      (res) => (options?.silent ? undefined : res.parsedBody),
+      (res) => (options?.silent ? undefined : res.parsedBody)
     );
   }
 
   saveAll(
     data: Array<documents.DocumentData<EntryInputType>>,
-    options?: documents.InsertDocumentOptions,
+    options?: documents.InsertDocumentOptions
   ) {
     return this._db.request(
       {
@@ -2631,14 +2658,14 @@ export class Collection<
         body: data,
         search: options,
       },
-      (res) => (options?.silent ? undefined : res.parsedBody),
+      (res) => (options?.silent ? undefined : res.parsedBody)
     );
   }
 
   replace(
     selector: documents.DocumentSelector,
     newData: documents.DocumentData<EntryInputType>,
-    options: documents.ReplaceDocumentOptions = {},
+    options: documents.ReplaceDocumentOptions = {}
   ) {
     const { ifMatch = undefined, ...opts } = options;
     const headers = {} as Record<string, string>;
@@ -2647,13 +2674,13 @@ export class Collection<
       {
         method: "PUT",
         pathname: `/_api/document/${encodeURI(
-          documents._documentHandle(selector, this._name),
+          documents._documentHandle(selector, this._name)
         )}`,
         headers,
         body: newData,
         search: opts,
       },
-      (res) => (options?.silent ? undefined : res.parsedBody),
+      (res) => (options?.silent ? undefined : res.parsedBody)
     );
   }
 
@@ -2662,7 +2689,7 @@ export class Collection<
       documents.DocumentData<EntryInputType> &
         ({ _key: string } | { _id: string })
     >,
-    options?: documents.ReplaceDocumentOptions,
+    options?: documents.ReplaceDocumentOptions
   ) {
     return this._db.request(
       {
@@ -2671,14 +2698,14 @@ export class Collection<
         body: newData,
         search: options,
       },
-      (res) => (options?.silent ? undefined : res.parsedBody),
+      (res) => (options?.silent ? undefined : res.parsedBody)
     );
   }
 
   update(
     selector: documents.DocumentSelector,
     newData: documents.Patch<documents.DocumentData<EntryInputType>>,
-    options: documents.UpdateDocumentOptions = {},
+    options: documents.UpdateDocumentOptions = {}
   ) {
     const { ifMatch = undefined, ...opts } = options;
     const headers = {} as Record<string, string>;
@@ -2687,13 +2714,13 @@ export class Collection<
       {
         method: "PATCH",
         pathname: `/_api/document/${encodeURI(
-          documents._documentHandle(selector, this._name),
+          documents._documentHandle(selector, this._name)
         )}`,
         headers,
         body: newData,
         search: opts,
       },
-      (res) => (options?.silent ? undefined : res.parsedBody),
+      (res) => (options?.silent ? undefined : res.parsedBody)
     );
   }
 
@@ -2702,7 +2729,7 @@ export class Collection<
       documents.Patch<documents.DocumentData<EntryInputType>> &
         ({ _key: string } | { _id: string })
     >,
-    options?: documents.UpdateDocumentOptions,
+    options?: documents.UpdateDocumentOptions
   ) {
     return this._db.request(
       {
@@ -2711,13 +2738,13 @@ export class Collection<
         body: newData,
         search: options,
       },
-      (res) => (options?.silent ? undefined : res.parsedBody),
+      (res) => (options?.silent ? undefined : res.parsedBody)
     );
   }
 
   remove(
     selector: documents.DocumentSelector,
-    options: documents.RemoveDocumentOptions = {},
+    options: documents.RemoveDocumentOptions = {}
   ) {
     const { ifMatch = undefined, ...opts } = options;
     const headers = {} as Record<string, string>;
@@ -2726,18 +2753,18 @@ export class Collection<
       {
         method: "DELETE",
         pathname: `/_api/document/${encodeURI(
-          documents._documentHandle(selector, this._name),
+          documents._documentHandle(selector, this._name)
         )}`,
         headers,
         search: opts,
       },
-      (res) => (options?.silent ? undefined : res.parsedBody),
+      (res) => (options?.silent ? undefined : res.parsedBody)
     );
   }
 
   removeAll(
     selectors: (string | documents.ObjectWithDocumentKey)[],
-    options?: documents.RemoveDocumentOptions,
+    options?: documents.RemoveDocumentOptions
   ) {
     return this._db.request(
       {
@@ -2746,7 +2773,7 @@ export class Collection<
         body: selectors,
         search: options,
       },
-      (res) => (options?.silent ? undefined : res.parsedBody),
+      (res) => (options?.silent ? undefined : res.parsedBody)
     );
   }
 
@@ -2754,7 +2781,7 @@ export class Collection<
     data: Buffer | Blob | string | any[],
     options: documents.ImportDocumentsOptions & {
       type?: "documents" | "list" | "auto";
-    } = {},
+    } = {}
   ): Promise<documents.ImportDocumentsResult> {
     const search = { ...options, collection: this._name };
     if (Array.isArray(data)) {
@@ -2776,7 +2803,7 @@ export class Collection<
   protected _edges(
     selector: documents.DocumentSelector,
     options: documents.DocumentEdgesOptions = {},
-    direction?: "in" | "out",
+    direction?: "in" | "out"
   ) {
     const { allowDirtyRead = undefined } = options;
     return this._db.request({
@@ -2791,21 +2818,21 @@ export class Collection<
 
   edges(
     vertex: documents.DocumentSelector,
-    options?: documents.DocumentEdgesOptions,
+    options?: documents.DocumentEdgesOptions
   ) {
     return this._edges(vertex, options);
   }
 
   inEdges(
     vertex: documents.DocumentSelector,
-    options?: documents.DocumentEdgesOptions,
+    options?: documents.DocumentEdgesOptions
   ) {
     return this._edges(vertex, options, "in");
   }
 
   outEdges(
     vertex: documents.DocumentSelector,
-    options?: documents.DocumentEdgesOptions,
+    options?: documents.DocumentEdgesOptions
   ) {
     return this._edges(vertex, options, "out");
   }
@@ -2817,10 +2844,10 @@ export class Collection<
       {
         method: "PUT",
         pathname: `/_api/collection/${encodeURIComponent(
-          this._name,
+          this._name
         )}/loadIndexesIntoMemory`,
       },
-      (res) => res.parsedBody.result,
+      (res) => res.parsedBody.result
     );
   }
 
@@ -2830,7 +2857,7 @@ export class Collection<
         pathname: "/_api/index",
         search: { collection: this._name, ...options },
       },
-      (res) => res.parsedBody.indexes,
+      (res) => res.parsedBody.indexes
     );
   }